split remote resource code into its own file
